requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.35
pymysql==1.1.0
pydantic-settings==2.2.1
python-dotenv==1.0.1
fastapi-pagination==0.12.25
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.1.0
alembic==1.16.5